decomposed image encoders
