+++ models/base_blocks.py
@@ -780,44 +780,3 @@ class CycleLR():
         x = np.abs(last_epoch / self.step_size - 2 * cycle + 1)
         lr_t = self.base_lr + (self.max_lr - self.base_lr) * np.maximum(0, (1-x)) * self.gamma**(last_epoch)
         return lr_t
-
-
-########################################    ConvNext   ######################################## 
-class ConvNext(nn.Module):
-    def __init__(self):
-        super().__init__()
-        base = timm.create_model("convnext_small_384_in22ft1k", pretrained=False)
-        self.base = nn.Sequential(*list(base.children())[:-1])
-
-    def forward(self,x):
-        x = self.base(x)
-        x = x.reshape(x.size(0), 98, -1)
-        return x  
-
-
-######################################## ResNet18      ######################################## 
-class ResNet18(nn.Module):
-    def __init__(self):
-        super().__init__()
-        base = models.resnet18(pretrained=False)
-        self.base = nn.Sequential(*list(base.children())[:-3])
-        in_features = base.fc.in_features
-        
-    def forward(self,x):
-        x = self.base(x)
-        x = x.view(x.size(0), 256, -1)
-        return x 
-
-
-########################################   ResNet50    ######################################## 
-class ResNet50(nn.Module):
-    def __init__(self):
-        super().__init__()
-        base = models.resnet50(pretrained=False)
-        self.base = nn.Sequential(*list(base.children())[:-1])
-        in_features = base.fc.in_features
-        
-    def forward(self,x):
-        x = self.base(x)
-        x = x.view(x.size(0), 8, -1)
-        return x
